New approach on the generation of the Map
New approach. Now the area of the map is divided into subareas and in each subarea a Room is generated and placed on the
map grid. The rooms can't intersect or touch by definition. The methods for the generation of the map have been modified
accordingly, some new method were written.

Still to find a solution to ensure an optimal split and the generation of the requested number of rooms. Right now, a
limit on the tries stops the Room generation loop, not very nice.