Wrote methods to handle the level change
Added the coordinates of the up stairs and the down stairs as fields to the Level class, to handle the positioning of
the player in the level switches.

The placeCharacter method has now different behaviours if the player is going one level up, one level down or just spawn
in the first level.

Bug fix on the Items spawn: replaced max with min.

The handleMovement method now has an infinite loop to handle the movement of the main character and returns a status
indicator, as an int.

The handleMovement now has a boolean to report if the player moved at all.