Modifications in the map generation
Remove the overlaps method.

Modified the connectToMap method to detect where the room is with respect to the picked point and to connect the point
to the graph accordingly.

Bug fix: the condition for the "empty space" in the itemsLayer and in the monstersLayer was wrong.

Added a method to return a single freeSpot in a Room. Modified the freeSpots method to use it.
Bug fix: the while has to go on while the spot is already in the set of freespots.

Added a method to detect wheter a particular location in the map is walkable or not.