Added the handling of the levels
Made some parameters to the level constructor parametric.

Added a loop to check the return status of handleMovement and to move the player one level up or down, and to create a
new level if needed.

Modified the methods for the printing of the map and the stats to update corretly with the new modifications.