Big bug fix in the map generation
Modified some methods to use less variables and removed useless ones.

Big bug fix: the wrong point was removed from the graph when connecting and disconnecting the point adjacent to the
picked point in the room. This way some nodes and edges in the graph where lost and, as a result, if a previously picked
point was picked again there would have been no paths to get to that point, it would be isolated from the graph. This
was fixed, now the rooms are generated correctly.