Major change to the structure
The monsters and items sets have been changed to maps, where the key is their id.

The monsterLayer and the upperLayer have been removed and replace with a single upperLayer, which now keeps the id of
the item/monster instead of the item/monster itself.

Many methods have been modified to reflect the changes, some have been removed, since they are now useless. The project
compiles and works fine.