Structural modifications
A upperLayer field has been added. It keeps the id of the item/monster on the tile, if present. All the constructors
have been modified accordingly. The get and set method have been written.

The constructors have been modified to utilize the other constructors in their initialization.

The isWalkablke method has been moved in the class.