Fixed free error, change in handleChoice
Fixed a memory corruption linked to the freeing of the menu (the order of memory freeing was wrong)

Change in handleChoice: the choice number is now a field in the class and is increased/decreased as you go up and down
the menu. The choice is now returned in constant time.