Modifications of the shopMenu method
Removed switch statement.

Added a boolean to check for failed purchases. This way the error message is correctly dislpayed.

Bug fix: a item would be assigned even if th emoney wasn't enough.

The name of the items is printed in the itemInfo window.

Rearranged the order of variable declaration/methods calling.