Major modifications
Modified the handleMovement method to take the bottom window as an argument.

Added the picking of objects and stepping on them if the inventory is full.

Added the battles to the gameplay.

Modified monstersAround to take the player position instead of the player itself.

Some bugfixes and corrections on the battle method.

The game is officially in beta.