Bug fix on level up and stats increase
Upon level up the def would be set to DEFMAX (i don't really understand why there needs to be a field for the max if the
max defence is a constant), "resetting" the increase due to the equipment.

Fixed the getters for luck and defence to return the min between the actual stat and the max.