Refactoring and polishing
Did some refactoring of the code: specifically i broke down the handleTurn method into a number of submethods with a
specific funtcion each.

Added some enum definition to make return statuses clearer.

Added a LevelConfig struct to have a lower number of arguments in the Level constructor.

Modified the involved code to reflect the new changes.

Bugfixes of the placeStairs and placeCharacter methods: now the stairs won't be in the way of an exit.

Made the writeInfo method a method member of the Level class.