Changes and bugfixes
Added a method that tells if a given point is inside a room. Makes the connectToMap method way simpler.

The generate method now returns the actual number of generated rooms, as they can be less than the expected number.

Removed the HALLWAY tile, replacing it with PAVEMENT.

Fixed and shortened the freeSpots method. It would try to find all the free spots in a single room if n > # of room, now
this has been fixed.